Encapsulate fork implementation
When forking a repository within the same forge, a name clash could
arise. To overcome this issue, we'll have to retry with a different,
unique name. The only way we can refer to this fork from then on is if
we store it's name in the database.

To make this possible, forge-specific operation will be defined in
interface.forges.base.Forge.fork_inner but users(callers) will use the
interface provided by `interface.git.fork`, which stores the fork's
name in the database.