feat: add HTML URL attribute to gitea_forge_repositories table
Mastodon requires AP actor required a HTML-accessible URL to redirect
users who'd rather see the Actor's posts in it's native environment. So
far, we were under the assumption that we could build repository URLs
from just owner and repository names. This works for GitHub family of
forges but not for other kinds of forges as their repository URL pattern
might be different